remove legacy neutral fill recipes
