+++ packages/web-components/fast-components/src/styles/recipes.ts
@@ -1,14 +1,6 @@
 import { cssCustomPropertyBehaviorFactory } from "@microsoft/fast-foundation";
 import {
-    neutralFill_DEPRECATED,
-    neutralFillToggle_DEPRECATED,
-    neutralFocus_DEPRECATED,
-    neutralFocusInnerAccent_DEPRECATED,
     neutralForeground_DEPRECATED,
-    neutralForegroundHint_DEPRECATED,
-    neutralForegroundHintLarge_DEPRECATED,
-    neutralForegroundToggle_DEPRECATED,
-    neutralForegroundToggleLarge_DEPRECATED,
     neutralLayerCard_DEPRECATED,
     neutralLayerCardContainer_DEPRECATED,
     neutralLayerFloating_DEPRECATED,
@@ -20,7 +12,6 @@ import {
     neutralOutline_DEPRECATED,
 } from "../color/index";
 import { FASTDesignSystemProvider } from "../design-system-provider/index";
-import { accentBaseColor } from "../fast-design-system";
 
 /**
  * Behavior to resolve and make available the neutral-foreground-rest CSS custom property.
@@ -60,138 +51,6 @@ export const neutralForegroundActiveBehavior = cssCustomPropertyBehaviorFactory(
 export const neutralForegroundFocusBehavior = cssCustomPropertyBehaviorFactory(
     "neutral-foreground-focus",
     x => neutralForeground_DEPRECATED(x).focus,
-    FASTDesignSystemProvider.findProvider
-);
-/**
- * Behavior to resolve and make available the neutral-foreground-toggle CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralForegroundToggleBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-foreground-toggle",
-    neutralForegroundToggle_DEPRECATED,
-    FASTDesignSystemProvider.findProvider
-);
-/**
- * Behavior to resolve and make available the neutral-foreground-toggle-large CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralForegroundToggleLargeBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-foreground-toggle-large",
-    neutralForegroundToggleLarge_DEPRECATED,
-    FASTDesignSystemProvider.findProvider
-);
-/**
- * Behavior to resolve and make available the neutral-foreground-hint CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralForegroundHintBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-foreground-hint",
-    neutralForegroundHint_DEPRECATED,
-    FASTDesignSystemProvider.findProvider
-);
-/**
- * Behavior to resolve and make available the neutral-foreground-hint-large CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralForegroundHintLargeBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-foreground-hint-large",
-    neutralForegroundHintLarge_DEPRECATED,
-    FASTDesignSystemProvider.findProvider
-);
-
-/**
- * Behavior to resolve and make available the neutral-fill-rest CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralFillRestBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-fill-rest",
-    x => neutralFill_DEPRECATED(x).rest,
-    FASTDesignSystemProvider.findProvider
-);
-/**
- * Behavior to resolve and make available the neutral-fill-hover CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralFillHoverBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-fill-hover",
-    x => neutralFill_DEPRECATED(x).hover,
-    FASTDesignSystemProvider.findProvider
-);
-/**
- * Behavior to resolve and make available the neutral-fill-active CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralFillActiveBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-fill-active",
-    x => neutralFill_DEPRECATED(x).active,
-    FASTDesignSystemProvider.findProvider
-);
-/**
- * Behavior to resolve and make available the neutral-fill-focus CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralFillFocusBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-fill-focus",
-    x => neutralFill_DEPRECATED(x).focus,
-    FASTDesignSystemProvider.findProvider
-);
-/**
- * Behavior to resolve and make available the neutral-fill-selected CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralFillSelectedBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-fill-selected",
-    x => neutralFill_DEPRECATED(x).selected,
-    FASTDesignSystemProvider.findProvider
-);
-
-/**
- * Behavior to resolve and make available the neutral-fill-toggle-rest CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralFillToggleRestBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-fill-toggle-rest",
-    x => neutralFillToggle_DEPRECATED(x).rest,
-    FASTDesignSystemProvider.findProvider
-);
-/**
- * Behavior to resolve and make available the neutral-fill-toggle-hover CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralFillToggleHoverBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-fill-toggle-hover",
-    x => neutralFillToggle_DEPRECATED(x).hover,
-    FASTDesignSystemProvider.findProvider
-);
-/**
- * Behavior to resolve and make available the neutral-fill-toggle-active CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralFillToggleActiveBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-fill-toggle-active",
-    x => neutralFillToggle_DEPRECATED(x).active,
-    FASTDesignSystemProvider.findProvider
-);
-/**
- * Behavior to resolve and make available the neutral-fill-toggle-focus CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralFillToggleFocusBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-fill-toggle-focus",
-    x => neutralFillToggle_DEPRECATED(x).focus,
     FASTDesignSystemProvider.findProvider
 );
 
@@ -314,25 +173,5 @@ export const neutralLayerL3Behavior = cssCustomPropertyBehaviorFactory(
 export const neutralLayerL4Behavior = cssCustomPropertyBehaviorFactory(
     "neutral-layer-l4",
     neutralLayerL4_DEPRECATED,
-    FASTDesignSystemProvider.findProvider
-);
-/**
- * Behavior to resolve and make available the neutral-focus CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralFocusBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-focus",
-    neutralFocus_DEPRECATED,
-    FASTDesignSystemProvider.findProvider
-);
-/**
- * Behavior to resolve and make available the neutral-focus-inner-accent CSS custom property.
- * @public
- * @deprecated - to-be deleted
- */
-export const neutralFocusInnerAccentBehavior = cssCustomPropertyBehaviorFactory(
-    "neutral-focus-inner-accent",
-    neutralFocusInnerAccent_DEPRECATED(accentBaseColor),
     FASTDesignSystemProvider.findProvider
 );
